refactor(data-dict): split test_extract_all.py to mirror extract/ package
